requirements.txt
# Python 3.11+ recommended for better performance
# Minimum: Python 3.9

sentence-transformers>=2.2.0
chromadb>=0.4.22
langchain-text-splitters>=0.0.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
numpy>=1.24.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
